pyproject.toml
[project]
name = "game-ai-workflow"
version = "0.1.0"
description = "AI Workflow for Game Server Business Code Generation"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "streamlit>=1.38.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "pandas>=2.2.0",
    "pydantic>=2.0.0",
    "tabulate>=0.9.0",
]

[project.scripts]
game-workflow = "app.main:main"
